Fixed compilation errors after having a big problem with circular dependencies.
