+++ src/GameManager.h
@@ -1,7 +1,6 @@
 #ifndef MY_PLAYDATE_TEST_GAME_MANAGER_H
 #define MY_PLAYDATE_TEST_GAME_MANAGER_H
 
-#include <string>
 #include "pdcpp/pdnewlib.h"
 #include "EntityManager.h"
 #include "Player.h"
