+++ src/Inventory.h
@@ -3,7 +3,7 @@
 
 #include <utility>
 #include <vector>
-#include "item.h"
+#include "Item.h"
 
 
 class Inventory
